refactor: move compiler errors to own file
