+++ src/compiler/mod.rs
@@ -1,1 +1,2 @@
 pub mod context;
+pub mod error;
